+++ CHANGELOG.md
@@ -3,6 +3,7 @@ All notable changes to this project will be documented in this file.
 
 ## [1.3.0] - 2021-03-28
 ### Changed
+- Add babel-eslint
 - Modify webpack.config.js for multiple locales
 - Add locale templates