+++ CHANGELOG.md
@@ -3,6 +3,7 @@ All notable changes to this project will be documented in this file.
 
 ## [1.3.0] - 2021-03-28
 ### Changed
+- Modify main template
 - Modify bundler config
 - Add babel
 - Modify webpack.config.js for multiple locales